Implement Phase 1b: /v1/messages/direct end-to-end
Routing core:
- internal/dispatch/strategy/strategy.go: RouteStrategy interface,
  RecipientHandle (UserID/ChatID/TopicID/Channel), Envelope (text +
  schema_version), Request, ResolveResult+ResolveError taxonomy
- internal/dispatch/strategy/direct.go: DirectStrategy + DirectResolver
  interface; PgDirectResolver resolves recipients against users +
  user_subscriptions + user_topics; per-recipient skip codes
  (unknown_recipient, recipient_inactive, recipient_not_subscribed,
  recipient_not_linked, recipient_topic_missing, bot_not_admin)
- internal/dispatch/dispatcher.go: Dispatcher interface + typed errors
  (ErrChatNotFound, ErrBotNotAdmin, ErrRateLimited, ErrTelegramAuth,
  ErrTransient)
- internal/dispatch/telegram/dispatcher.go: telego-backed Dispatcher;
  classify() maps Telegram error strings to typed errors; respects
  chat-grain limiter (waits or returns ErrRateLimited)
- internal/dispatch/telegram/dispatch_limiter.go: two-tier token bucket
  (global 25/s burst 30 + per-chat 1/s burst 2) wrapping our
  RateLimiter interface for the dispatch side

HTTP surface:
- internal/api/handlers/messages_direct.go: POST /v1/messages/direct
  validates envelope.schema_version (must be 1; 0 -> missing, 99 ->
  unsupported), enforces messages.direct.send capability, runs
  strategy.Resolve, emits audit chain
  received -> validated -> {dispatched -> delivered|deferred|denied}
  with delivery_channel='supergroup'; per-recipient outcome report
- removed internal/api/handlers/noop.go (Phase 1a scaffolding gone)
- internal/api/server.go: wired DirectHandler with capability gate
- internal/api/middleware/request_id.go: exported WithTraceID for tests
- cmd/server/main.go: telego.NewBot + DispatchLimiter + DirectStrategy
  wired into Deps

Audit writer fix:
- internal/audit/writer.go: cast NULLIF placeholders for bigint columns
  ($7/$11/$12 -> ::bigint) so pgx infers bigint, not int4. Without the
  cast pgx encodes recipient_chat_id=-1001234567890 as int4 and the
  INSERT silently fails (rows for stages dispatched/deferred dropped).
  Verified by retrying smoke and observing 4-stage audit chain.
- audit_write_failed is now logged (was silently swallowed)

Migration:
- migrations/0003_seed_dev_fixture_user: one fully-linked dev user
  (id 1, telegram_id 100000042, personal_supergroup -1001234567890,
  bot_is_admin true) subscribed to app deploy-alerts with topic_id 7;
  drives Phase 1b smoke

Tests (all pass):
- direct strategy: empty app/recipients rejection, resolver forwarding,
  error propagation, name() stability
- direct handler: happy path emits 4-stage chain to stub audit;
  missing/unsupported envelope version; empty recipients;
  recipient_not_subscribed skip path; dispatcher error -> deferred +
  classifyDispatchErr; strategy error -> 500

Phase 1b smoke (against docker compose):
- /v1/messages/direct happy:
  - request: app_id=deploy-alerts, recipient=user 1, text+schema_version=1
  - response 200, per-recipient failed=1 reason='telegram_auth_failed'
    (placeholder bot token rejected by Telegram - this IS the expected
     classification on a non-real token; perimeter + resolver +
     dispatcher + audit all worked)
  - audit_log: 4 rows trace_id=dbg-003
      received -> validated ->
      dispatched (channel='supergroup',uid=1,cid=-1001234567890) ->
      deferred (error='telegram_auth_failed')
- denied: dev-user lacking messages.direct.send -> 403 forbidden
- skipped: unknown recipient 9999 -> 200 skipped=1 'unknown_recipient'
- skipped: recipient not subscribed -> 200 'recipient_not_subscribed'
- 400: missing envelope.schema_version
- 400: unsupported envelope.schema_version=99
- secret leak: 0 occurrences of bearer cleartext or bot token in stdout

docker compose:
- postgres host port moved 5432 -> 5433 (5432 conflicted locally)
- TELEGRAM_BOT_TOKEN default now matches telego regex ^\d+:[\w-]{35}$
  (1:AAAA... 35 chars) so the app starts without a real token

+++ internal/audit/writer.go
@@ -27,8 +27,9 @@ func (w *PgWriter) Write(ctx context.Context, e Event) error {
 			recipient_user_id, recipient_chat_id, error_code, details_json
 		) VALUES (
 			COALESCE($1, now()), NULLIF($2,''), NULLIF($3,''), $4, NULLIF($5,''),
-			NULLIF($6,''), NULLIF($7,0), NULLIF($8,''), NULLIF($9,''),
-			NULLIF($10,''), NULLIF($11,0), NULLIF($12,0), NULLIF($13,''), $14
+			NULLIF($6,''), NULLIF($7::bigint, 0), NULLIF($8,''), NULLIF($9,''),
+			NULLIF($10,''), NULLIF($11::bigint, 0), NULLIF($12::bigint, 0),
+			NULLIF($13,''), $14
 		)`
 	at := e.At
 	var atArg any
